adding html and css structure
