+++ .gitignore
@@ -1,1 +1,2 @@
 .Ds_Store
+.log
