+++ templates/index.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html lang="en, fr">
+
+
+  <head>
+
+        <meta charset="UTF-8" />
+        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>CloudAcademy</title>
+        <script src="https://kit.fontawesome.com/3ee5796760.js" crossorigin="anonymous"></script>
+        <link rel="stylesheet" href="../static/style.css" type="text/css" />
+        <!-- <link rel="stylesheet" href="{{ url_for('static', filename='style.css')}}"> -->
+        <link rel="preconnect" href="https://fonts.gstatic.com"> 
+        <link rel="preconnect" href="https://fonts.gstatic.com"> 
+        <link href="https://fonts.googleapis.com/css2?family=Raleway:wght@200&family=Roboto:wght@300&display=swap" rel="stylesheet">
+
+    </head>
+
+
+  <body>
+
+      <nav id="navbar">
+        
+
+            <div class="nav_extemity">
+
+                <li><h1>CloudAcademy</h1></li>
+            </div>
+
+            <div class="container_nav">
+                <!-- <button><i class="fas fa-search"></i></button> -->
+                <li><a href="#home" class="nav_link">Home</a></li>
+                <li><a href="#feature"class="nav_link">Feature</a></li>
+                <div>
+                    <input type="text" class="search" placeholder="Choose a theme"/>
+                    <a href="" class="nav_link"><i class="fas search fa-search"></i></a>
+                </div> 
+
+            </div>
+
+            <div class="nav_extemity">
+                <li><a href="#faq" class="nav_link button-bar">F-A-Q</a></li>
+            </div>
+
+        
+    </nav>
+
+      <main class="content">
+
+
+        <section class="info">
+            
+            <div class="container sloggan" id="home">
+                <h1>Want to learn how to be an amazing cloud developer ? 
+                    Here is the place to be !</h1>
+            </div>  
+
+            <div id="icon">
+                <i class="main_icon fab fa-python"></i></i>
+                <i class="main_icon fab fa-js-square"></i>
+                <i class="main_icon fab fa-docker"></i>
+            </div>
+            
+        </section>
+
+        <div class="main_container">
+
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+          <div class="video_link"></div>
+
+        </div>
+      </main>
+
+
+    </div>
+  </body>
+</html>
